requirements.txt
kafka-python
transformers
torch
cloudevents
influxdb
influxdb-client
requests
